Making setup file more verbose about what's being done.
The setup.sh file is meant to provide messages to the user of what it's
doing at the time. I seem to have forgotten to add the echo message for
when it copies tinyfugue-scripts/*.tf to /usr/local/share/tf-lib/ which
is something that I meant to do when first setting up the script. I just
now noticed that it wasn't added, so it's been fixed now.